+++ apps/docs/app/page.tsx
@@ -2,9 +2,9 @@ import { ComingSoonButton, Header } from "@rubriclab/ui";
 
 export default function Page() {
   return (
-    <>
+    <div className="flex items-center flex-col justify-center">
       <Header text="Docs" />
       <ComingSoonButton />
-    </>
+    </div>
   );
 }