+++ packages/@rubriclab/typescript-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rubriclab/typescript-config
 
+## 0.1.0
+
+### Minor Changes
+
+- minor release all packages
+
 ## 0.0.3
 
 ### Patch Changes