+++ packages/@rubriclab/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rubriclab/ui
 
+## 0.1.0
+
+### Minor Changes
+
+- minor release all packages
+
 ## 0.0.6
 
 ### Patch Changes